New render command + IBL cubemap prefiltering

+++ Iteration12/Engine/FFT.cpp
@@ -4,9 +4,9 @@
 #include "fft.h"
 #include <stdint.h>
 
-FFT::FFT(unsigned int N) : N(N), reversed(0), W(0), pi2(2 * PI) 
+FFT::FFT(unsigned int N) : N(N), reversed(0), W(0), pi2(2.0f * PI) 
 {
-	log_2_N = logf(N)/logf(2);
+	log_2_N = (unsigned int)(logf(N)/logf(2));
 	parity = 0;
 
 	// Prepare bit-reserved indices
@@ -149,7 +149,7 @@ Complex FFT::TwiddleFactor(UINT x, UINT N)
 
 void FFT::IterativeFFT( Complex* primal, Complex* dual, UINT stride, UINT zeroOffset )
 {
-	const int absP = logf(N)/logf(2.0f);
+	const int absP = (int)logf(N)/logf(2.0f);
 
 	// Bottom level of iteration tree
 	for (int i = 0; i < N; i++)
@@ -160,7 +160,7 @@ void FFT::IterativeFFT( Complex* primal, Complex* dual, UINT stride, UINT zeroOf
 	{
 		// Complex root of unity
 		const int unityStep = 0x1 << p;
-		const double theta = 2.0f * PI / unityStep; // INVERSE
+		const float theta = 2.0f * PI / unityStep; // INVERSE
 		const Complex unityRoot(cos(theta), sin(theta));
 
 		// Each higher level doubles the step size
@@ -192,7 +192,7 @@ void FFT::IterativeFFT( Complex* primal, Complex* dual, UINT stride, UINT zeroOf
 
 void FFT::ParallelFFT( UINT index, Complex* primal, Complex* dual, UINT stride, UINT zeroOffset )
 {
-	const int absP = logf(N)/logf(2.0f);
+	const int absP = (int)logf(N)/logf(2.0f);
 
 	// Bottom level of iteration tree
 	for (int i = 0; i < N; i++)
@@ -203,7 +203,7 @@ void FFT::ParallelFFT( UINT index, Complex* primal, Complex* dual, UINT stride, 
 	{
 		// Complex root of unity
 		const int unityStep = 0x1 << p;
-		const double theta = 2.0f * PI / unityStep; // INVERSE
+		const float theta = 2.0f * PI / unityStep; // INVERSE
 		const Complex unityRoot(cos(theta), sin(theta));
 
 		// Each higher level doubles the step size
